requirements-dev.txt
-r requirements.txt
coverage
django-nose
flake8
mock
nose
tox
